added instagram links to profiles
