+++ src/App.css
@@ -40,3 +40,11 @@
 .playerprofile{
   margin: auto;
 }
+.profile-social{
+  display: none;
+}
+
+.player-card:hover .profile-social {
+  display:block;
+}
+
